Fix autosave handling
